+++ R/s_pattern.R
@@ -4,7 +4,7 @@
 #'
 #'
 #' The \link[=str_arithmetic]{%s-%, %s/%, %ss%} operators,
-#' as well as the string detection operators (\link{str_truth}),
+#' as well as the string search operators (\link{str_truth}),
 #' perform pattern matching for some purpose,
 #' where the pattern is given on the right hand side. \cr
 #' When a character vector or string is given on the right hand side,